requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=4.2.0=py39haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
async-lru=2.0.4=py39haa95532_0
attrs=23.1.0=py39haa95532_0
babel=2.11.0=py39haa95532_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=py39haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.7=py39h9128911_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
ca-certificates=2024.2.2=h56e8100_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.16.0=py39h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=py39haa95532_0
comm=0.1.2=py39haa95532_0
contourpy=1.2.0=py39h59b6b97_0
cuda-cccl=12.3.101=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.3.101=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py39hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
exceptiongroup=1.2.0=py39haa95532_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.13.1=py39haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
gmpy2=2.1.2=py39h7f96b67_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
idna=3.4=py39haa95532_0
importlib-metadata=7.0.1=py39haa95532_0
importlib_metadata=7.0.1=hd3eb1b0_0
importlib_resources=6.1.1=py39haa95532_1
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.28.0=py39haa95532_0
ipython=8.15.0=py39haa95532_0
jedi=0.18.1=py39haa95532_1
jinja2=3.1.3=py39haa95532_0
joblib=1.2.0=py39haa95532_0
jpeg=9e=h2bbff1b_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.19.2=py39haa95532_0
jsonschema-specifications=2023.7.1=py39haa95532_0
jupyter-lsp=2.2.0=py39haa95532_0
jupyter_client=8.6.0=py39haa95532_0
jupyter_core=5.5.0=py39haa95532_0
jupyter_events=0.8.0=py39haa95532_0
jupyter_server=2.10.0=py39haa95532_0
jupyter_server_terminals=0.4.4=py39haa95532_1
jupyterlab=4.0.11=py39haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.25.1=py39haa95532_0
kiwisolver=1.4.4=py39hd77b12b_0
krb5=1.20.1=h5b6d351_0
lame=3.100=hcfcfb64_1003
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.4.107=0
libcurand-dev=10.3.4.107=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_1
libflac=1.4.3=h63175ca_0
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libogg=1.3.4=h8ffe710_1
libopus=1.3.1=h8ffe710_1
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libsndfile=1.2.2=h81429f1_1
libsodium=1.0.18=h62dcd97_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=h0e60522_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.3=py39h2bbff1b_0
matplotlib=3.8.0=py39haa95532_0
matplotlib-base=3.8.0=py39h4ed8f06_0
matplotlib-inline=0.1.6=py39haa95532_0
mistune=2.0.4=py39haa95532_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.8=py39h2bbff1b_0
mkl_random=1.2.4=py39h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpg123=1.32.4=h63175ca_0
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py39haa95532_0
munkres=1.1.4=py_0
nbclient=0.8.0=py39haa95532_0
nbconvert=7.10.0=py39haa95532_0
nbformat=5.9.2=py39haa95532_0
nest-asyncio=1.5.6=py39haa95532_0
networkx=3.1=py39haa95532_0
notebook-shim=0.2.3=py39haa95532_0
numexpr=2.8.7=py39h2cd9be0_0
numpy=1.26.3=py39h055cbcc_0
numpy-base=1.26.3=py39h65a83cf_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.2.1=hcfcfb64_0
overrides=7.4.0=py39haa95532_0
packaging=23.1=py39haa95532_0
pandas=2.1.4=py39h4ed8f06_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.2.0=py39h2bbff1b_0
pip=23.3.1=py39haa95532_0
platformdirs=3.10.0=py39haa95532_0
ply=3.11=py39haa95532_0
prometheus_client=0.14.1=py39haa95532_0
prompt-toolkit=3.0.43=py39haa95532_0
psutil=5.9.0=py39h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py39haa95532_1
pyparsing=3.0.9=py39haa95532_0
pyqt=5.15.10=py39hd77b12b_0
pyqt5-sip=12.13.0=py39h2bbff1b_0
pysoundfile=0.12.1=pypyhd8ed1ab_1
python=3.9.18=h1aa4202_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py39haa95532_0
python-json-logger=2.0.7=py39haa95532_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.2.0=py3.9_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py39haa95532_0
pywin32=305=py39h2bbff1b_0
pywinpty=2.0.10=py39h5da7b33_0
pyyaml=6.0.1=py39h2bbff1b_0
pyzmq=25.1.2=py39hd77b12b_0
qt-main=5.15.2=h19c9488_10
referencing=0.30.2=py39haa95532_0
requests=2.31.0=py39haa95532_1
rfc3339-validator=0.1.4=py39haa95532_0
rfc3986-validator=0.1.1=py39haa95532_0
rpds-py=0.10.6=py39h062c2fa_0
scikit-learn=1.1.3=py39hd77b12b_0
scipy=1.11.4=py39h309d312_0
seaborn=0.12.2=py39haa95532_0
send2trash=1.8.2=py39haa95532_0
setuptools=68.2.2=py39haa95532_0
sip=6.7.12=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sklearn-pandas=2.2.0=py39haa95532_0
sniffio=1.3.0=py39haa95532_0
soupsieve=2.5=py39haa95532_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.12=py39haa95532_0
tbb=2021.8.0=h59b6b97_0
terminado=0.17.1=py39haa95532_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=py39haa95532_0
tomli=2.0.1=py39haa95532_0
torchaudio=2.2.0=pypi_0
torchvision=0.17.0=pypi_0
tornado=6.3.3=py39h2bbff1b_0
traitlets=5.7.1=py39haa95532_0
typing-extensions=4.9.0=py39haa95532_1
typing_extensions=4.9.0=py39haa95532_1
tzdata=2023d=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.1.0=py39haa95532_0
vc=14.2=h21ff451_1
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39haa95532_1
websocket-client=0.58.0=py39haa95532_4
wfdb=4.1.2=pyhd8ed1ab_0
wheel=0.41.2=py39haa95532_0
winpty=0.4.3=4
xz=5.4.5=h8cc25b3_0
yaml=0.2.5=he774522_0
zeromq=4.3.5=hd77b12b_0
zipp=3.17.0=py39haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
